+++ Assets/Scripts/UiScripts/SoundManager.cs
@@ -3,95 +3,106 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
+[System.Serializable]
+public class Sound 
+{
+    public string soundName;
+    public AudioClip clip;
+}
+
 public class SoundManager : MonoBehaviour
 {
-    private static SoundManager instance;
-
-    private AudioSource audioSource;
-
-
-    private GameObject[] musics;
-
     public GameObject BgmMixer;
+    public GameObject SfxMixer;
 
-
-
-    public static SoundManager Instance
+    private static SoundManager instance;
+    public static SoundManager Instance 
     {
-        get
-        {
-            return instance == null ? null : instance;
-
-        }
+        get { return instance; }
     }
-    public AudioSource AudioSource
-    {
-        get
-        {
-            return audioSource == null ? null : audioSource;
-        }
 
-    }
+    [Header("사운드 등록")]
+    [SerializeField] Sound[] bgmSounds;
+    [SerializeField] Sound[] sfxSounds;
+
+    [Header("브금 플레이어")]
+    [SerializeField] AudioSource bgmPlayer;
+
+    [Header("효과음 플레이어")]
+    [SerializeField] AudioSource[] sfxPlayer;
+
+    private Dictionary<string, AudioClip> soundDictionary = new Dictionary<string, AudioClip>();
+
     private void Awake()
     {
         if (instance == null)
         {
             instance = this;
+            DontDestroyOnLoad(gameObject);
         }
         else 
         {
-            Destroy(this.gameObject);
+            Destroy(gameObject);
         }
 
-        musics = GameObject.FindGameObjectsWithTag("Music");
-        if (musics.Length >= 2)
+        foreach (Sound sound in bgmSounds)
         {
-            Destroy(this.gameObject);
+            soundDictionary[sound.soundName] = sound.clip;
         }
 
-        DontDestroyOnLoad(transform.gameObject);
-        audioSource = GetComponent<AudioSource>();
+        BgmMixer.GetComponent<Slider>().onValueChanged.AddListener(SetBgmVolume);
+        SfxMixer.GetComponent<Slider>().onValueChanged.AddListener(setSfxVolume);
+    }
+    void Start()
+    {
+        PlayBGM("BackgroundMusic");
 
-        BgmMixer.GetComponent<Slider>().onValueChanged.AddListener(SetVolume);
     }
 
-    public void PlayMusic()
+    public void PlayBGM(string soundName)
     {
-        if (audioSource.isPlaying)
+        if (soundDictionary.ContainsKey(soundName))
         {
-            return;
+            AudioClip soundClip = soundDictionary[soundName];
+            bgmPlayer.clip = soundClip;
+            bgmPlayer.Play();
         }
-        audioSource.Play();
+        else
+        {
+            Debug.LogWarning("사운드를 찾을 수 없음: " + soundName);
+        }
     }
 
-    public void StopMusic()
+    public void PlaySE(string _soundName) 
     {
-        audioSource.Stop();
-    }
+        for (int i = 0; i < sfxSounds.Length; i++) 
+        {
+            if (_soundName == sfxSounds[i].soundName) 
+            {
+                for (int j = 0; j < sfxPlayer.Length; j++) 
+                {
+                    if (!sfxPlayer[j].isPlaying) 
+                    {
+                        sfxPlayer[j].clip = sfxSounds[i].clip;
+                        sfxPlayer[j].Play();
 
-    public void SetVolume(float value)
+                        return;
+                    }
+                }
+                return;
+            }
+        }
+    }
+    public void SetBgmVolume(float value)
     {
-        audioSource.volume = value;
+        bgmPlayer.volume = value;
+    }
+    public void setSfxVolume(float value) 
+    {
+        foreach (AudioSource player in sfxPlayer)
+        {
+            player.volume = value;
+        }
     }
 
-
-
-
-
-
-    //private bool isPaused = false; 배경음 사운드 퍼즈
-
-    //private void Update()
-    //{
-    //    if (Time.timeScale == 0 && !isPaused)
-    //    {
-    //        audioSource.Pause();
-    //        isPaused = true;
-    //    }
-    //    else if(Time.timeScale !=0 &&isPaused)
-    //    {
-    //        audioSource.UnPause();
-    //        isPaused = false;
-    //    }
-    //}
 }
