+++ Assets/Scripts/UiScripts/SoundManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8f31d9057537ccc49b68814734d16541
+guid: 62ae96a7e4a73cb49b8660d0d064b45c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
