--- Assets/Scripts/UiScripts/TestSoundManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 62ae96a7e4a73cb49b8660d0d064b45c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
